+++ src/games/snake/entities/background/components/BackgroundDisplayable.hpp
@@ -76,8 +76,8 @@ class BackgroundDisplayable : public shared::games::components::ITextureComponen
          * @brief Get the Position object
          * @return Vector2i&
          */
-        Vector2i &getPosition(void) noexcept override;
-        Vector2i _position;
+        Vector2f &getPosition(void) noexcept override;
+        Vector2f _position;
 
     private:
         const shared::games::entity::IEntity &_entity;
